fifth commit
The program determines whether one is eligible to vote based on their age and citizenship. One should be 18 years of age or older and a citizen of a country in East Africa.

The program also caters to the problem of cases typed by the user. It converts all the user input cases to lowercase using the .lower() function